split MethodConfigSpec
